check for node alive
