+++ src/plugin.ts
@@ -16,7 +16,7 @@ import {
   removePendingSession,
 } from './data-service/pending-sessions-service';
 import {
-  allocateDeviceForSession,
+  allocateDeviceForSession, checkNodeServerAvailability,
   cronReleaseBlockedDevices,
   deviceType,
   initlializeStorage,
@@ -29,10 +29,8 @@ import { Container } from 'typedi';
 import logger from './logger';
 import { v4 as uuidv4 } from 'uuid';
 import axios from 'axios';
-import { hubUrl, isHub, stripAppiumPrefixes } from './helpers';
-
+import {hubUrl, isHub, spinWith, stripAppiumPrefixes} from './helpers';
 import { addProxyHandler, registerProxyMiddlware } from './wd-command-proxy';
-import ora from 'ora';
 import ChromeDriverManager from './device-managers/ChromeDriverManager';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
@@ -40,7 +38,6 @@ import { addCLIArgs, getCLIArgs } from './data-service/pluginArgs';
 import Cloud from './enums/Cloud';
 import ip from 'ip';
 import _ from 'lodash';
-import asyncWait from 'async-wait-until';
 
 const commandsQueueGuard = new AsyncLock();
 const DEVICE_MANAGER_LOCK_NAME = 'DeviceManager';
@@ -97,13 +94,17 @@ class DevicePlugin extends BasePlugin {
     logger.info(
       `📣📣📣 Device Farm Plugin will be served at 🔗 http://localhost:${cliArgs.port}/device-farm`
     );
-    if (isHub(cliArgs)) await DevicePlugin.waitForRemoteHubServerToBeRunning(cliArgs);
+    if (isHub(cliArgs)) {
+      const hub = cliArgs.plugin['device-farm'].hub;
+      await DevicePlugin.waitForRemoteHubServerToBeRunning(hub);
+    }
     const devicesUpdates = await updateDeviceList(cliArgs);
     if (isIOS(cliArgs) && deviceType(cliArgs, 'simulated')) {
       await setSimulatorState(devicesUpdates);
       await refreshSimulatorState(cliArgs);
     }
     await cronReleaseBlockedDevices();
+    await checkNodeServerAvailability();
   }
 
   private static setIncludeSimulatorState(cliArgs: any, deviceTypes: string) {
@@ -115,14 +116,13 @@ class DevicePlugin extends BasePlugin {
     return deviceTypes;
   }
 
-  private static async waitForRemoteHubServerToBeRunning(cliArgs: any) {
-    const hub = cliArgs.plugin['device-farm'].hub;
+  static async waitForRemoteHubServerToBeRunning(host: string) {
     await spinWith(
-      `Waiting for node server ${hub} to be up and running\n`,
+      `Waiting for node server ${host} to be up and running\n`,
       async () => {
         await axios({
           method: 'get',
-          url: `${hub}/device-farm`,
+          url: `${host}/device-farm`,
           timeout: 30000,
           headers: {
             'Content-Type': 'application/json',
@@ -236,28 +236,6 @@ class DevicePlugin extends BasePlugin {
     logger.info(`📱 Unblocking the device that is blocked for session ${sessionId}`);
     return await next();
   }
-}
-
-// eslint-disable-next-line @typescript-eslint/no-empty-function
-async function spinWith(msg: string, fn: () => any, callback = (msg: string) => {}) {
-  const spinner = ora(msg).start();
-  await asyncWait(
-    async () => {
-      try {
-        await fn();
-        spinner.succeed();
-        return true;
-      } catch (err) {
-        spinner.fail();
-        if (callback) callback(msg);
-        return false;
-      }
-    },
-    {
-      intervalBetweenAttempts: 2000,
-      timeout: 60 * 1000,
-    }
-  );
 }
 
 Object.assign(DevicePlugin.prototype, commands);
